requirements_TEST.txt
colorlog==6.9.0
homeassistant==2025.1.4
pytest-homeassistant-custom-component==0.13.205 # Matching version for 2025.1.4
pip>=21.3.1
ruff==0.9.1
mypy==1.4.1
dvg-pid-controller==2.2.0
pytest-asyncio
